refactor: add CarName class
